use u_int16_t instead of u_int8_t

+++ tests/kana/KanaTest.cpp
@@ -133,7 +133,7 @@ TEST(KanaTest, CheckHiragana) {
   auto& sourceMap{Kana::getMap(CharType::Hiragana)};
   EXPECT_EQ(sourceMap.size(), TotalKana);
   // count various types including smallDigraphs (which should be 0)
-  u_int8_t hanDakutenMonographs{}, smallMonographs{}, plainMonographs{},
+  u_int16_t hanDakutenMonographs{}, smallMonographs{}, plainMonographs{},
       dakutenMonographs{}, plainDigraphs{}, hanDakutenDigraphs{},
       dakutenDigraphs{}, smallDigraphs{};
   for (auto& i : sourceMap) {
@@ -240,7 +240,7 @@ TEST(KanaTest, CheckKatakana) {
 TEST(KanaTest, CheckRomaji) {
   auto& sourceMap{Kana::getMap(CharType::Romaji)};
   EXPECT_EQ(sourceMap.size(), TotalRomaji);
-  u_int8_t aNum{}, vaNum{}, iNum{}, viNum{}, uNum{}, vuNum{}, eNum{}, veNum{},
+  u_int16_t aNum{}, vaNum{}, iNum{}, viNum{}, uNum{}, vuNum{}, eNum{}, veNum{},
       oNum{}, voNum{}, nNum{};
   std::set<std::string> romajiVariants;
   for (auto& i : sourceMap) {
